// src/app/shared/services/api/model/car-response.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CarResponse { 
    id?: number;
    year?: number;
    mileageCounter?: number;
    seatingCapacity?: number;
    plate?: string;
    color?: string;
    price?: number;
    state?: number;
    description?: string;
    modelId?: number;
    transmissionTypeId?: number;
    fuelTypeId?: number;
}